scripts/requirements.txt
numpy>=1.24
pillow>=10
scipy>=1.11
